Separate RaidBossesRequest into {,All}RaidBossesRequest
