+++ protocol/src/main/protobuf/walfie/gbf/raidfinder/protocol/requests.proto
@@ -11,14 +11,21 @@ option (scalapb.options) = {
 
 message RequestMessage {
   oneof data {
-    RaidBossesRequest raidBossesMessage = 1;
-    FollowRequest followMessage = 2;
-    UnfollowRequest unfollowMessage = 3;
+    AllRaidBossesRequest allRaidBossesMessage = 1;
+    RaidBossesRequest raidBossesMessage = 2;
+    FollowRequest followMessage = 3;
+    UnfollowRequest unfollowMessage = 4;
   };
+};
+
+message AllRaidBossesRequest {
+  option (scalapb.message).extends = "Request";
 };
 
 message RaidBossesRequest {
   option (scalapb.message).extends = "Request";
+
+  repeated string bossNames = 1;
 };
 
 message FollowRequest {
